+++ apps/mobile/app/(student)/(tabs)/classrooms.tsx
@@ -1,5 +1,7 @@
 import { StudentClassroomsScreen } from "../../../src/student-foundation"
+import { useDoubleBackToExit } from "../../../src/use-back-handler"
 
 export default function StudentClassroomsRoute() {
+  useDoubleBackToExit()
   return <StudentClassroomsScreen />
 }