+++ apps/mobile/app/(teacher)/(tabs)/classrooms.tsx
@@ -1,5 +1,7 @@
 import { TeacherClassroomsScreen } from "../../../src/teacher-foundation"
+import { useDoubleBackToExit } from "../../../src/use-back-handler"
 
 export default function TeacherClassroomsRoute() {
+  useDoubleBackToExit()
   return <TeacherClassroomsScreen />
 }